game over and game won menu screen logic added
handled from gameManager which is now a non-persistent mono singleton.

Changed MonoGenericSingleton  back to being non-persistent(removed dontDestroyOnLoad) as it works better for the specific usecase of restarting the game, instead of having to reassign references if i was using a generic mono singleton, which itself isnt a true singleton.